requirements.txt
# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Azure Document Intelligence SDK (v4)
azure-ai-documentintelligence==1.0.0b4

# Configuration
jinja2==3.1.3
pydantic-settings==2.2.1

# CSV/Excel export
openpyxl==3.1.2

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
